+++ contracts/PKPNFT.sol
@@ -53,12 +53,12 @@ contract PKPNFT is
     /* ========== VIEWS ========== */
 
     /// get the eth address for the keypair, as long as it's an ecdsa keypair
-    function getEthAddress(uint256 tokenId) public view returns (address) {
+    function getEthAddress(uint tokenId) public view returns (address) {
         return router.getEthAddress(tokenId);
     }
 
     /// includes the 0x04 prefix so you can pass this directly to ethers.utils.computeAddress
-    function getPubkey(uint256 tokenId) public view returns (bytes memory) {
+    function getPubkey(uint tokenId) public view returns (bytes memory) {
         return router.getPubkey(tokenId);
     }
 
@@ -135,7 +135,7 @@ contract PKPNFT is
             );
     }
 
-    function exists(uint256 tokenId) public view returns (bool) {
+    function exists(uint tokenId) public view returns (bool) {
         return _exists(tokenId);
     }
