+++ contracts/PKPNFTMetadata.sol
@@ -33,7 +33,7 @@ contract PKPNFTMetadata {
 
         bytes memory _base = "0123456789abcdef";
 
-        for (uint256 i = 0; i < buffer.length; i++) {
+        for (uint i = 0; i < buffer.length; i++) {
             converted[i * 2] = _base[uint8(buffer[i]) / _base.length];
             converted[i * 2 + 1] = _base[uint8(buffer[i]) % _base.length];
         }
